testing some widgets2

+++ src/ConcreteEngine.Graphics/Gfx/Resources/GfxResourceApi.cs
@@ -18,7 +18,7 @@ public sealed class GfxResourceApi
     public nint GetNativeHandle<TId>(TId id) where TId : unmanaged, IResourceId
     {
         var handle = _storeHub.GetHandleStore<TId>().GetHandle(id);
-        return (nint)_backendHub.GetStore(handle.Kind).GetNativeHandle(handle).Value;
+        return _backendHub.GetStore(handle.Kind).GetNativeHandle(handle).Value;
     }
 
     public TMeta GetMeta<TId, TMeta>(TId id) where TId : unmanaged, IResourceId where TMeta : unmanaged, IResourceMeta
